requirements.txt
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.19.2
matplotlib>=3.3.2
tensorboard>=2.4.0
tqdm>=4.51.0
scikit-learn>=0.23.2
pillow>=8.0.1 